+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## **1.0.6** 2018-03-12
+
+### Added
+
+- String extension and unit tests.
+- Timer extension and unit tests.
+
 ## **1.0.5** 2018-03-12
 
 ### Added